Flatten statuses, polls, timelines, search, and context into Hunter
