+++ lib/hunter/poll.ex
@@ -20,7 +20,6 @@ defmodule Hunter.Poll do
       (only with user token)
 
   """
-  alias Hunter.Api.HTTPClient
 
   @type t :: %__MODULE__{
           id: String.t(),
@@ -48,35 +47,4 @@ defmodule Hunter.Poll do
     :voted,
     :own_votes
   ]
-
-  @doc """
-  Retrieve a poll
-
-  ## Parameters
-
-    * `conn` - connection credentials
-    * `id` - poll identifier
-
-  """
-  @spec poll(Hunter.Client.t(), String.t() | non_neg_integer) :: Hunter.Poll.t()
-  def poll(conn, id) do
-    HTTPClient.poll(conn, id)
-  end
-
-  @doc """
-  Vote on one or more options in a poll
-
-  ## Parameters
-
-    * `conn` - connection credentials
-    * `id` - poll identifier
-    * `choices` - list of option indices to vote for (zero-based); multiple
-      choices are only allowed on multiple-choice polls
-
-  """
-  @spec vote(Hunter.Client.t(), String.t() | non_neg_integer, [non_neg_integer]) ::
-          Hunter.Poll.t()
-  def vote(conn, id, choices) do
-    HTTPClient.vote(conn, id, choices)
-  end
 end
