+++ lib/hunter/status.ex
@@ -47,7 +47,6 @@ defmodule Hunter.Status do
   **NOTE**: When `spoiler_text` is present, `sensitive` is true
 
   """
-  alias Hunter.Api.HTTPClient
 
   @type t :: %__MODULE__{
           id: String.t(),
@@ -85,8 +84,6 @@ defmodule Hunter.Status do
           quote_approval: map | nil
         }
 
-  @type status_id :: String.t() | non_neg_integer
-
   @derive [Poison.Encoder]
   defstruct [
     :id,
@@ -123,444 +120,4 @@ defmodule Hunter.Status do
     :quote,
     :quote_approval
   ]
-
-  @doc """
-  Create new status
-
-  ## Parameters
-
-    * `conn` - connection credentials
-    * `status` - text of the status
-    * `options` - option list
-
-  ## Options
-
-    * `in_reply_to_id` - local ID of the status you want to reply to
-    * `media_ids` - list of media IDs to attach to the status (maximum: 4)
-    * `sensitive` - whether the media of the status is NSFW
-    * `spoiler_text` - text to be shown as a warning before the actual content
-    * `visibility` - either `direct`, `private`, `unlisted` or `public`
-    * `language` - ISO 639-1 language code for the status
-    * `poll` - map with `options` (list of choices) and `expires_in`
-      (seconds); optional: `multiple`, `hide_totals`
-    * `quoted_status_id` - ID of a status to quote
-    * `scheduled_at` - ISO 8601 datetime (at least 5 minutes ahead) at which
-      the status should be published; the call returns a
-      `Hunter.ScheduledStatus` instead of a `Hunter.Status`
-    * `idempotency_key` - unique string sent as the `Idempotency-Key` header
-      to prevent duplicate submissions
-
-  """
-  @spec create_status(Hunter.Client.t(), String.t(), Keyword.t()) ::
-          Hunter.Status.t() | Hunter.ScheduledStatus.t() | no_return
-  def create_status(conn, status, options \\ []) do
-    HTTPClient.create_status(conn, status, options)
-  end
-
-  @doc """
-  Retrieve status
-
-  ## Parameters
-
-    * `conn` - connection credentials
-    * `id` - status identifier
-
-  """
-  @spec status(Hunter.Client.t(), status_id) :: Hunter.Status.t()
-  def status(conn, id) do
-    HTTPClient.status(conn, id)
-  end
-
-  @doc """
-  Retrieve multiple statuses by id
-
-  ## Parameters
-
-    * `conn` - connection credentials
-    * `ids` - list of status identifiers
-
-  """
-  @spec statuses_by_ids(Hunter.Client.t(), [status_id]) :: [Hunter.Status.t()]
-  def statuses_by_ids(conn, ids) do
-    HTTPClient.statuses_by_ids(conn, ids)
-  end
-
-  @doc """
-  Edit a status
-
-  ## Parameters
-
-    * `conn` - connection credentials
-    * `id` - status identifier
-    * `status` - the new text of the status
-    * `options` - option list
-
-  ## Options
-
-    * `spoiler_text` - text to be shown as a warning before the actual content
-    * `sensitive` - whether the media of the status is NSFW
-    * `language` - ISO 639-1 language code for the status
-    * `media_ids` - list of media IDs to attach to the status
-    * `poll` - map with `options` (list of choices) and `expires_in`
-      (seconds); replaces the current poll
-
-  """
-  @spec edit_status(Hunter.Client.t(), status_id, String.t(), Keyword.t()) ::
-          Hunter.Status.t() | no_return
-  def edit_status(conn, id, status, options \\ []) do
-    HTTPClient.edit_status(conn, id, status, options)
-  end
-
-  @doc """
-  Retrieve the edit history of a status
-
-  ## Parameters
-
-    * `conn` - connection credentials
-    * `id` - status identifier
-
-  """
-  @spec status_history(Hunter.Client.t(), status_id) :: [Hunter.StatusEdit.t()]
-  def status_history(conn, id) do
-    HTTPClient.status_history(conn, id)
-  end
-
-  @doc """
-  Retrieve the plain-text source of a status, for editing
-
-  ## Parameters
-
-    * `conn` - connection credentials
-    * `id` - status identifier
-
-  """
-  @spec status_source(Hunter.Client.t(), status_id) :: Hunter.StatusSource.t()
-  def status_source(conn, id) do
-    HTTPClient.status_source(conn, id)
-  end
-
-  @doc """
-  Bookmark a status
-
-  ## Parameters
-
-    * `conn` - connection credentials
-    * `id` - status identifier
-
-  """
-  @spec bookmark(Hunter.Client.t(), status_id) :: Hunter.Status.t()
-  def bookmark(conn, id) do
-    HTTPClient.bookmark(conn, id)
-  end
-
-  @doc """
-  Remove a status from bookmarks
-
-  ## Parameters
-
-    * `conn` - connection credentials
-    * `id` - status identifier
-
-  """
-  @spec unbookmark(Hunter.Client.t(), status_id) :: Hunter.Status.t()
-  def unbookmark(conn, id) do
-    HTTPClient.unbookmark(conn, id)
-  end
-
-  @doc """
-  Fetch the user's bookmarked statuses
-
-  ## Parameters
-
-    * `conn` - connection credentials
-    * `options` - option list
-
-  ## Options
-
-    * `max_id` - get a list of bookmarks with id less than or equal this value
-    * `since_id` - get a list of bookmarks with id greater than this value
-    * `min_id` - get a list of bookmarks with id greater than this value,
-      immediately newer
-    * `limit` - maximum number of bookmarks to get, default: 20, max: 40
-
-  """
-  @spec bookmarks(Hunter.Client.t(), Keyword.t()) :: [Hunter.Status.t()]
-  def bookmarks(conn, options \\ []) do
-    HTTPClient.bookmarks(conn, options)
-  end
-
-  @doc """
-  Pin a status to the top of the user's profile
-
-  ## Parameters
-
-    * `conn` - connection credentials
-    * `id` - status identifier
-
-  """
-  @spec pin(Hunter.Client.t(), status_id) :: Hunter.Status.t()
-  def pin(conn, id) do
-    HTTPClient.pin(conn, id)
-  end
-
-  @doc """
-  Unpin a status from the user's profile
-
-  ## Parameters
-
-    * `conn` - connection credentials
-    * `id` - status identifier
-
-  """
-  @spec unpin(Hunter.Client.t(), status_id) :: Hunter.Status.t()
-  def unpin(conn, id) do
-    HTTPClient.unpin(conn, id)
-  end
-
-  @doc """
-  Mute a conversation, so its thread stops generating notifications
-
-  ## Parameters
-
-    * `conn` - connection credentials
-    * `id` - status identifier
-
-  """
-  @spec mute_conversation(Hunter.Client.t(), status_id) :: Hunter.Status.t()
-  def mute_conversation(conn, id) do
-    HTTPClient.mute_conversation(conn, id)
-  end
-
-  @doc """
-  Unmute a conversation
-
-  ## Parameters
-
-    * `conn` - connection credentials
-    * `id` - status identifier
-
-  """
-  @spec unmute_conversation(Hunter.Client.t(), status_id) :: Hunter.Status.t()
-  def unmute_conversation(conn, id) do
-    HTTPClient.unmute_conversation(conn, id)
-  end
-
-  @doc """
-  Translate a status into some language
-
-  ## Parameters
-
-    * `conn` - connection credentials
-    * `id` - status identifier
-    * `options` - option list
-
-  ## Options
-
-    * `lang` - ISO 639-1 language code the status should be translated into;
-      defaults to the user's current locale
-
-  """
-  @spec translate_status(Hunter.Client.t(), status_id, Keyword.t()) :: Hunter.Translation.t()
-  def translate_status(conn, id, options \\ []) do
-    HTTPClient.translate_status(conn, id, options)
-  end
-
-  @doc """
-  Destroy status
-
-  ## Parameters
-
-    * `conn` - connection credentials
-    * `id` - status identifier
-
-  """
-  @spec destroy_status(Hunter.Client.t(), status_id) :: boolean
-  def destroy_status(conn, id) do
-    HTTPClient.destroy_status(conn, id)
-  end
-
-  @doc """
-  Reblog a status
-
-  ## Parameters
-
-    * `conn` - connection credentials
-    * `id` - status identifier
-
-  """
-  @spec reblog(Hunter.Client.t(), status_id) :: Hunter.Status.t()
-  def reblog(conn, id) do
-    HTTPClient.reblog(conn, id)
-  end
-
-  @doc """
-  Undo a reblog of a status
-
-  ## Parameters
-
-  * `conn` - connection credentials
-  * `id` - status identifier
-
-  """
-  @spec unreblog(Hunter.Client.t(), status_id) :: Hunter.Status.t()
-  def unreblog(conn, id) do
-    HTTPClient.unreblog(conn, id)
-  end
-
-  @doc """
-  Favorite a status
-
-  ## Parameters
-
-    * `conn` - connection credentials
-    * `id` - status identifier
-
-  """
-  @spec favourite(Hunter.Client.t(), status_id) :: Hunter.Status.t()
-  def favourite(conn, id) do
-    HTTPClient.favourite(conn, id)
-  end
-
-  @doc """
-  Undo a favorite of a status
-
-  ## Parameters
-
-    * `conn` - connection credentials
-    * `id` - status identifier
-
-  """
-  @spec unfavourite(Hunter.Client.t(), status_id) :: Hunter.Status.t()
-  def unfavourite(conn, id) do
-    HTTPClient.unfavourite(conn, id)
-  end
-
-  @doc """
-  Fetch a user's favourites
-
-  ## Parameters
-
-    * `conn` - connection credentials
-    * `options` - option list
-
-  ## Options
-
-    * `max_id` - get a list of favourites with id less than or equal this value
-    * `since_id` - get a list of favourites with id greater than this value
-    * `limit` - maximum of favourites to get, default: 20, max: 40
-
-  """
-  @spec favourites(Hunter.Client.t(), Keyword.t()) :: [Hunter.Status.t()]
-  def favourites(conn, options \\ []) do
-    HTTPClient.favourites(conn, options)
-  end
-
-  @doc """
-  Get a list of statuses by a user
-
-  ## Parameters
-
-    * `conn` - connection credentials
-    * `account_id` - account identifier
-    * `options` - option list
-
-  ## Options
-
-    * `only_media` - only return `Hunter.Status.t` that have media attachments
-    * `exclude_replies` - skip statuses that reply to other statuses
-    * `max_id` - get a list of statuses with id less than or equal this value
-    * `since_id` - get a list of statuses with id greater than this value
-    * `limit` - maximum number of statuses to get, default: 20, max: 40
-
-  """
-  @spec statuses(Hunter.Client.t(), status_id, Keyword.t()) :: [Hunter.Status.t()]
-  def statuses(conn, account_id, options \\ []) do
-    HTTPClient.statuses(conn, account_id, Map.new(options))
-  end
-
-  @doc """
-  Retrieve statuses from the home timeline
-
-  ## Parameters
-
-    * `conn` - connection credentials
-    * `options` - option list
-
-  ## Options
-
-    * `max_id` - get a list of timelines with id less than or equal this value
-    * `since_id` - get a list of timelines with id greater than this value
-    * `limit` - maximum number of statuses on the requested timeline to get, default: 20, max: 40
-
-  """
-  @spec home_timeline(Hunter.Client.t(), Keyword.t()) :: [Hunter.Status.t()]
-  def home_timeline(conn, options \\ []) do
-    HTTPClient.home_timeline(conn, Map.new(options))
-  end
-
-  @doc """
-  Retrieve statuses from the public timeline
-
-  ## Parameters
-
-    * `conn` - connection credentials
-    * `options` - option list
-
-  ## Options
-
-    * `local` - only return statuses originating from this instance
-    * `max_id` - get a list of timelines with id less than or equal this value
-    * `since_id` - get a list of timelines with id greater than this value
-    * `limit` - maximum number of statuses on the requested timeline to get, default: 20, max: 40
-
-  """
-  @spec public_timeline(Hunter.Client.t(), Keyword.t()) :: [Hunter.Status.t()]
-  def public_timeline(conn, options \\ []) do
-    HTTPClient.public_timeline(conn, Map.new(options))
-  end
-
-  @doc """
-  Retrieve statuses from a hashtag
-
-  ## Parameters
-
-    * `conn` - connection credentials
-    * `hashtag` - string list
-    * `options` - option list
-
-  ## Options
-
-    * `local` - only return statuses originating from this instance
-    * `max_id` - get a list of timelines with id less than or equal this value
-    * `since_id` - get a list of timelines with id greater than this value
-    * `limit` - maximum number of statuses on the requested timeline to get, default: 20, max: 40
-
-  """
-  @spec hashtag_timeline(Hunter.Client.t(), [String.t()], Keyword.t()) :: [Hunter.Status.t()]
-  def hashtag_timeline(conn, hashtag, options \\ []) do
-    HTTPClient.hashtag_timeline(conn, hashtag, Map.new(options))
-  end
-
-  @doc """
-  Retrieve statuses from the given list's timeline
-
-  ## Parameters
-
-    * `conn` - connection credentials
-    * `list_id` - list identifier
-    * `options` - option list
-
-  ## Options
-
-    * `max_id` - get a list of timelines with id less than or equal this value
-    * `since_id` - get a list of timelines with id greater than this value
-    * `limit` - maximum number of statuses on the requested timeline to get, default: 20, max: 40
-
-  """
-  @spec list_timeline(Hunter.Client.t(), String.t() | non_neg_integer, Keyword.t()) :: [
-          Hunter.Status.t()
-        ]
-  def list_timeline(conn, list_id, options \\ []) do
-    HTTPClient.list_timeline(conn, list_id, Map.new(options))
-  end
 end
